Move data files
