+++ static/app/views/performance/http/httpSamplesPanel.tsx
@@ -26,12 +26,12 @@ import {normalizeUrl} from 'sentry/utils/withDomainRequired';
 import {AverageValueMarkLine} from 'sentry/views/performance/charts/averageValueMarkLine';
 import {DurationChart} from 'sentry/views/performance/http/charts/durationChart';
 import {ResponseCodeCountChart} from 'sentry/views/performance/http/charts/responseCodeCountChart';
-import {HTTP_RESPONSE_STATUS_CODES} from 'sentry/views/performance/http/definitions';
+import {HTTP_RESPONSE_STATUS_CODES} from 'sentry/views/performance/http/data/definitions';
+import {useSpanSamples} from 'sentry/views/performance/http/data/useSpanSamples';
 import decodePanel from 'sentry/views/performance/http/queryParameterDecoders/panel';
 import decodeResponseCodeClass from 'sentry/views/performance/http/queryParameterDecoders/responseCodeClass';
 import {SpanSamplesTable} from 'sentry/views/performance/http/tables/spanSamplesTable';
 import {useDebouncedState} from 'sentry/views/performance/http/useDebouncedState';
-import {useSpanSamples} from 'sentry/views/performance/http/useSpanSamples';
 import {MetricReadout} from 'sentry/views/performance/metricReadout';
 import * as ModuleLayout from 'sentry/views/performance/moduleLayout';
 import {computeAxisMax} from 'sentry/views/starfish/components/chart';
